Change to conversion using converter interface
Implement escape sequence processing as a converter.
